Mostly fixed wavefunction calculation
Wavefunctions are now calculated in a manner that seems reasonable and
consistent with theory.  Previous calculations had wavefunctions that
would change scale drastically between nearby energies.  This no longer
happens.

const.m has a new method added that calculates the conversion between
an energy in K to an inverse wavenumber given the reduced mass of the
system

For speed (maybe), the calculation of the intermediate region of the Rb
BO potential has been replaced with a direct sum rather than a for loop.
I have not checked how much this might change the calculation speed.