Re-factored code to be more object-oriented
Major revision of the code base to be more object oriented and also to
make the code a little more sensible.  Removed references to anonymous
functions for calling potential functions and now use a single class
method.

Hamiltonian generation is done inside the atompair and atomproperties
classes.

Multichannel.m now is a single function.  Integration via the improved
log-derivative method of Manolopoulos is now handled in a separate
function.  Its variable list is not compatible with the one for bound
states because the grid is generated at each call.

The class alkalioperators should be useable with bound state calculations